PackageManagerTest: Directly use the synthetic projects tree for testing
As this test only cares about detecting the presence of definition files
(and not about the declared dependencies in detail) it is fine to rely on
the files even if their content might change from time to time.

This change is a preparation for skipping pathologic cases of empty
definition files, which otherwise would cause parsing errors with several
third-party libraries that we use, like Jackson.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch.io>